Added feature

+++ OOP_Project_6/accounts.java
@@ -55,7 +55,8 @@ public class accounts {
         }
 
         // CHECKS IF CONDITIONS ARE MET
-        if (password.equals(passwordmatcher) && password.length() >= 8 && exist == false) {
+        if (password.equals(passwordmatcher) && password.length() >= 8 && exist == false
+                && password.matches("^[a-zA-Z0-9]+$") && username.matches("^[a-zA-Z0-9._]+$")) {
             try {
                 FileWriter records = new FileWriter(
                         "/home/daniel/Documents/03_programming_projects/Java_OOP_Projects/OOP_Project_6/records.txt",
@@ -74,10 +75,14 @@ public class accounts {
 
         } else if (exist == true) {
             System.out.println("\nUsername is not available");
+        } else if (!username.equals("^[a-zA-Z0-9._]+$")) {
+            System.out.println("\nUsername cannot contain spaces");
         } else if (!password.equals(passwordmatcher)) {
             System.out.println("\nSorry, the password you entered does not match");
         } else if (password.length() < 8) {
             System.out.println("\nPassword must be 8 or more characters");
+        } else if (!password.matches("^[a-zA-Z0-9]+$")) {
+            System.out.println("\nPassword cannot contain symbols");
         }
         sc.close();
         System.out.println(bar);
